Add MIMXRT1010 target
I acquired the MIMXRT1011 CMSIS Pack from ARM, and used the target-gen
tool to generate something that looked like this target file. I then
hand edited that output into the form you see today.

The main issue with the stock Pack is that the NVM memory map targets
ROMCP, and not the FlexSPI memory region. I don't know if that's
actually what's in the FLM, or if there was a parsing issue. I touched
this up to represent the 16MiB flash region suggested by the flashing
algorithm.

The Pack output has two MIMXRT1011 variants. They were the same, so I
consolidated them into a single MIMXRT1010 variant. I made the numbers
hex, and added human-readable names to the memory regions. Flash
algorithm was untouched (except for cosmetic number changes).

Although this was generated from an MIMXRT1011 Pack, I changed the name
to MIMXRT1010 (1011 vs 1010, part vs family). pyOCD uses the 1010 name
instead of 1011, so I figured I'd keep it consistent.