Fix flash algorithm memory range selection
The check as written fails if the NVM range is larger than the flash
algorithm's range, though that's an OK situation.  On the other hand,
the check passes if the NVM range is smaller than the flash algorithm's
range. That's no good, since it could let us write beyond the NVM range
if the target were poorly specified.

This commit flips the check. Now, we ensure the NVM range contains the
flash algorithm's range. Tested on an MIMXRT1010 system, where the NVM
region is larger (504MiB) than the flash algorithm's region (16MiB).