Correct the NVM region
The 1176 DFP describes the ROMCP region as the readable-executable
memory region. ROMCP isn't NVM. I'm not sure why the DFP uses this.

This is an issue in the DFP description file. We're unlikely to correct
this in target-gen.

This is the first issue I'd hit when trying to flash the target. It
manifests like

  cargo run --package=probe-rs-cli -- download --chip=mimxrt1170 ...
  ...
  Error: Error while flashing

  Caused by:
      No flash memory contains the entire requested memory range 0x30000000..0x30002000.

To fix it, either look at the memory map in the reference manual, or
observe what memory the mimxrt117x_quadspi_4kb_sec flash algorithm wants
to work with.