Prefer FlexRAM boot fuse config for MIMXRT MCUs
Our MIMXRT target definitions expect to place the flash algorithm into
an available OCRAM region, typically the second OCRAM FlexRAM bank. If
the user's program changes the FlexRAM bank configuration at runtime,
that configuration persists across reset. If that configuration took
away the OCRAM FlexRAM bank that we're using for the flash programming
algorithm, we won't be able to place the flash algorithm. This manifests
as an error early during programming.

This commit updates the debug sequences to force the boot fuse FlexRAM
configuration. This represents the POR state of the MCU, so user
programs already need to handle this out of reset. There's still a
gotchya: if the user changed the boot fuses to remove the OCRAM bank
we're targeting, this isn't going to help. On the other hand, the
manufacturer-shipped boot fuse configuration will always work, so we're
hoping that's the common case.

Tested on a 1010EVK and an 1170EVK.