Add user-controlled MIMXRT FlexRAM config
The MIMXRT10xx and 11xx debug sequences check for an environment
variable containing a FlexRAM bank layout specification. If not set, the
debug sequence resets FlexRAM to the boot fuse values, as usual. But if
set and correct, the debug sequence will write the configuration to the
MCU before resetting. The value persists after reset.

Users who are RAM loading their firmware can set this environment
variable to control FlexRAM layout before probe-rs loads the firmware
into RAM.

Quick demo:

    PROBE_RS_MIMXRT_FLEXRAM_LAYOUT=OODI probe-rs run ...

After reset, banks 0 and 1 are OCRAM (O). Bank 2 is DTCM (D), and bank 3
is ITCM (I). All unspecified banks are unused (U), so this is the same
as a longer

    PROBE_RS_MIMXRT_FLEXRAM_LAYOUT=OODIUUUUUUUUUUUU probe-rs run ...

The maximum number of banks in a 10xx and 11xx chip is 16. However, some
chips have fewer banks. probe-rs doesn't check any chip-specific limits,
so you'll need to do this.

To maintain backwards compatibility, do not define the environment
variable. A defined but empty environment variable sets all banks to
unused, as if you wrote 'U' for the spec.

If the spec is invalid, the debug sequence bails out of the reset
sequence. It might be nice to detect an invalid spec earlier, but I
haven't figured out where that could go.